# Fix crash when there isn't enough data